requirements.txt
Django==5.2.8
djangorestframework==3.16.1
celery==5.5.3
redis==7.1.0
psycopg2-binary==2.9.11
pydantic==2.12.4
fastapi==0.121.3
uvicorn==0.38.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
billiard==4.2.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
h11==0.16.0
idna==3.11
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.50.0
transformers==4.46.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
cryptography==42.0.5
prometheus-client==0.20.0
